completed

+++ controllerFunctions/orderFunc.js
@@ -1,19 +1,20 @@
 const orderModel = require("../models/orderModel");
-const mongoose = require('mongoose')
+const mongoose = require("mongoose");
+const moment = require("moment");
 
 // Get all orders
 const getAllOrders = async (req, res) => {
   const orders = await orderModel.find();
 
-  return res.json({ status: true, orders });
+  return res.status(200).json({ status: true, orders });
 };
 // Get a single order
-const getSingleOrder = async(req, res)=>{
+const getSingleOrder = async (req, res) => {
   const { orderId } = req.params;
-  if(!mongoose.Types.ObjectId.isValid(orderId)){
-    return res.status(404).json({error: 'No such Id'})
+  if (!mongoose.Types.ObjectId.isValid(orderId)) {
+    return res.status(404).json({ error: "No such Id" });
   }
-  
+
   const order = await orderModel.findById(orderId);
 
   if (!order) {
@@ -21,7 +22,7 @@ const getSingleOrder = async(req, res)=>{
   }
 
   return res.json({ status: true, order });
-}
+};
 
 // Post/Create orders
 const createOrder = async (req, res) => {
@@ -31,23 +32,25 @@ const createOrder = async (req, res) => {
     prev += curr.price;
     return prev;
   }, 0);
-  try{
-
+  try {
     const order = await orderModel.create({
       items: body.items,
       created_at: moment().toDate(),
       total_price,
     });
-  
+
     return res.json({ status: true, order });
-  }catch(error){
-    res.status(400).json({error: error.message})
+  } catch (error) {
+    res.status(400).json({ error: error.message });
   }
 };
 // Update Order
 const updateOrder = async (req, res) => {
   const { id } = req.params;
   const { state } = req.body;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such Id" });
+  }
 
   const order = await orderModel.findById(id);
 
@@ -71,6 +74,9 @@ const updateOrder = async (req, res) => {
 // Delete order
 const deleteOrder = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such Id" });
+  }
 
   const order = await orderModel.deleteOne({ _id: id });
 
@@ -82,5 +88,5 @@ module.exports = {
   getAllOrders,
   getSingleOrder,
   updateOrder,
-  deleteOrder
+  deleteOrder,
 };
